Support compiling outside of a 'workspace'
Some language server clients - notably, vim with the ALE plugin - do not
actually support the concept of workspaces. What that client (and
presumably others that do not support workspaces) do is send over a
rootUri (or if they're very old, a rootPath) that describes the root of
*the* workspace.

This patch enables ink-language-server to support such clients - if they
do not explicitly indicate workspace support in the initialization
method, but do set a rootUri or rootPath - we assume that the path sent
by the client is the root of a singular workspace.